# doc/documentation/swaggerApi.json
{
  "swagger": "2.0",
  "info": {
    "title": "清球汇",
    "version": "last"
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "身份验证"
        ],
        "summary": "登录",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "thuId": {
                  "type": "string",
                  "description": "清华证件号",
                  "default": ""
                },
                "password": {
                  "type": "string",
                  "description": "密码"
                },
                "rememberMe": {
                  "type": "boolean",
                  "description": "是否七天内登陆有效"
                }
              },
              "required": [
                "thuId",
                "password",
                "rememberMe"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "成功的额外信息，若成功则为\"ok\"。"
                },
                "userId": {
                  "type": "string",
                  "description": "用户 ID"
                }
              },
              "required": [
                "message",
                "userId"
              ]
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "身份验证"
        ],
        "summary": "注册账户",
        "description": "目前还不清楚能否通过清华的授权码，向清华方面请求用户的学号等数据。因此，此处默认学号等隐私信息的获取由服务端完成，故注册方面只上传其他必须信息。\n**当前测试阶段，\"code\"字段即为学号**！",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "username": {
                  "type": "string",
                  "description": "用户昵称"
                },
                "password": {
                  "type": "string",
                  "description": "密码"
                },
                "ticket": {
                  "type": "string",
                  "description": "授权码"
                },
                "email": {
                  "type": "string",
                  "description": "可选邮箱"
                },
                "mobile": {
                  "type": "string",
                  "description": "可选11位手机号"
                }
              },
              "required": [
                "username",
                "password",
                "ticket"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "message"
              ]
            }
          }
        }
      }
    },
...
    "/user/info": {
      "get": {
        "tags": [
          "用户信息"
        ],
        "summary": "个人信息",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "gender",
                "message",
                "role",
                "thuId",
                "userId",
                "username",
                "createdAt"
              ],
              "properties": {
                "avatar": {
                  "type": "string",
                  "format": "url",
                  "description": "用户头像"
                },
                "email": {
                  "type": "string",
                  "description": "邮箱"
                },
                "gender": {
                  "type": "string",
                  "description": "性别",
                  "enum": [
                    "FEMALE",
                    "MALE",
                    "OTHER",
                    "UNKNOWN"
                  ]
                },
                "message": {
                  "type": "string",
                  "description": "简短成功讯息"
                },
                "mobile": {
                  "type": "string",
                  "description": "手机"
                },
                "role": {
                  "type": "string",
                  "description": "用户权限",
                  "enum": [
                    "ADMIN",
                    "ROOT",
                    "USER"
                  ]
                },
                "thuId": {
                  "type": "string",
                  "description": "清华证件 ID"
                },
                "userId": {
                  "type": "string",
                  "description": "用户 ID"
                },
                "username": {
                  "type": "string",
                  "description": "用户名"
                },
                "createdAt": {
                  "type": "string",
                  "description": "创建时间",
                  "format": "date-time"
                }
              },
              "title": "UserInfoResponse",
              "$$ref": "#/definitions/UserInfoResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "用户信息"
        ],
        "summary": "修改个人信息",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "avatar": {
                  "type": "string",
                  "description": "头像的文件名(key)",
                  "pattern": "^AVATAR-[a-zA-Z0-9.-]+$"
                },
                "email": {
                  "type": "string",
                  "description": "邮箱"
                },
                "gender": {
                  "type": "string",
                  "description": "性别",
                  "enum": [
                    "FEMALE",
                    "MALE",
                    "OTHER",
                    "UNKNOWN"
                  ]
                },
                "mobile": {
                  "type": "string",
                  "description": "手机",
                  "pattern": "^\\d{11}$"
                },
                "newPassword": {
                  "type": "string",
                  "description": "新密码"
                },
                "oldPassword": {
                  "type": "string",
                  "description": "旧密码"
                },
                "username": {
                  "type": "string",
                  "description": "用户名"
                }
              },
              "title": "UserUpdateRequest",
              "$$ref": "#/definitions/UserUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "message"
              ]
            }
          }
        }
      }
    },
    "/user/check": {
      "get": {
        "tags": [
          "用户信息"
        ],
        "summary": "按用户名检查用户存在性",
        "description": "如果找不到用户，返回的 userId 是 null",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "用户名",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "ok"
                },
                "userId": {
                  "type": "string",
                  "description": "可能是 null"
                }
              },
              "required": [
                "message"
              ]
            }
          }
        }
      }
    },
...
  }
}